requirements.txt
requests>=2.28.0
python-nmap>=0.7.1
wafw00f>=2.2.0
reportlab>=3.6.12
beautifulsoup4>=4.11.0
cryptography>=39.0.0
dnspython>=2.3.0
urllib3>=1.26.0
lxml>=4.9.0